Conflicts:
	src/Basescape/BaseInfoState.cpp
	src/Basescape/BasescapeState.cpp
	src/Basescape/CraftEquipmentState.cpp
	src/Basescape/CraftEquipmentState.h
	src/Basescape/ManageAlienContainmentState.cpp
	src/Basescape/ManageAlienContainmentState.h
	src/Basescape/MiniBaseView.cpp
	src/Basescape/MiniBaseView.h
	src/Basescape/PurchaseState.cpp
	src/Basescape/PurchaseState.h
	src/Basescape/SellState.cpp
	src/Basescape/SellState.h
	src/Basescape/TransferItemsState.cpp
	src/Basescape/TransferItemsState.h
	src/Battlescape/Projectile.h
	src/Battlescape/TileEngine.cpp
	src/Geoscape/AllocatePsiTrainingState.h
	src/Geoscape/Globe.cpp
	src/Geoscape/GraphsState.cpp
	src/Geoscape/GraphsState.h
	src/Interface/TextEdit.cpp
	src/Interface/TextList.cpp
	src/Interface/TextList.h
	src/Menu/OptionsAdvancedState.cpp
	src/Menu/OptionsModsState.cpp
	src/Menu/SavedGameState.cpp
	src/Ruleset/RuleCountry.cpp
	src/Ruleset/RuleRegion.cpp
	src/Savegame/Craft.cpp
	src/Ufopaedia/ArticleStateItem.cpp

+++ src/Basescape/BaseInfoState.cpp
@@ -186,7 +186,7 @@ BaseInfoState::BaseInfoState(
 	_mini->setTexture(_game->getResourcePack()->getSurfaceSet("BASEBITS.PCK"));
 	_mini->setBases(_game->getSavedGame()->getBases());
 
-	for (unsigned
+	for (size_t
 			i = 0;
 			i < _game->getSavedGame()->getBases()->size();
 			++i)
@@ -445,7 +445,7 @@ void BaseInfoState::edtBaseChange(Action* action)
  */
 void BaseInfoState::miniClick(Action*)
 {
-	unsigned int base = _mini->getHoveredBase();
+	size_t base = _mini->getHoveredBase();
 	if (base < _game->getSavedGame()->getBases()->size())
 	{
 		_mini->setSelectedBase(base);
