+++ src/Basescape/BasescapeState.cpp
@@ -142,7 +142,7 @@ BasescapeState::BasescapeState(
 
 	_mini->setTexture(_game->getResourcePack()->getSurfaceSet("BASEBITS.PCK"));
 	_mini->setBases(_game->getSavedGame()->getBases());
-	for (unsigned
+	for (size_t
 			i = 0;
 			i < _game->getSavedGame()->getBases()->size();
 			++i)
@@ -359,7 +359,7 @@ void BasescapeState::setBase(Base* base)
 {
 	_base = base;
 
-	for (unsigned int
+	for (size_t
 			i = 0;
 			i < _game->getSavedGame()->getBases()->size();
 			++i)
@@ -672,7 +672,7 @@ void BasescapeState::viewMouseOut(Action*)
  */
 void BasescapeState::miniClick(Action*)
 {
-	unsigned base = _mini->getHoveredBase();
+	size_t base = _mini->getHoveredBase();
 	if (base < _game->getSavedGame()->getBases()->size())
 	{
 		_mini->setSelectedBase(base);
