+++ src/Basescape/MiniBaseView.cpp
@@ -92,7 +92,7 @@ void MiniBaseView::setTexture(SurfaceSet* texture)
  * Returns the base the mouse cursor is currently over.
  * @return ID of the base.
  */
-unsigned int MiniBaseView::getHoveredBase() const
+size_t MiniBaseView::getHoveredBase() const
 {
 	return _hoverBase;
 }
@@ -118,7 +118,7 @@ void MiniBaseView::draw()
 {
 	Surface::draw();
 
-	for (unsigned int
+	for (size_t
 			i = 0;
 			i < MAX_BASES;
 			++i)
